# openapi.yaml
openapi: 3.0.3
info:
  title: Signature Device API - OpenAPI 3.0
  description: |-
    REST API documentation for the Signature Device API
  contact:
    name: Giacomo Cortesi
    url: https://giacomocortesi.github.io/hugo-gc
    email: giacomo.cortesi1993@gmail.com
  version: 1.0.0
servers:
- url: http://{username}:{port}/api/v0
  description: API server
  variables:
    username:
      default: localhost
    port:
      default: '8080'

tags:
  - name: devices
    description: Signature Devices

paths:
  /devices:
    get:
      summary: List all signature devices
      description: Retrieves a list of all available signature devices.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SignatureDeviceResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new signature device
      description: Creates a new signature device.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignatureDeviceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignatureDeviceResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /devices/{id}:
    get:
      summary: Get a signature device by ID
      description: Retrieves a specific signature device by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignatureDeviceResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /devices/{id}/signatures:
    get:
      summary: Get all signatures for a signature device
      description: Retrieves a list of all signatures generated by the specified signature device.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SignatureResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Sign transaction data using a signature device
      description: Signs the provided transaction data using the specified signature device.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: string
                  description: Transaction data to be signed.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignatureResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /health:
    get:
      summary: Checks the health of the service
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  version:
                    type: string
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: string                             
components:
  schemas:
    SignatureDeviceRequest:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the signature device (optional), if not specified a random one is pick by the server
        algorithm:
          type: string
          description: Signature algorithm used by the device
          enum:
            - RSA
            - ECC
        label:
          type: string
          description: Human-readable label for the device (optional)
    SignatureDeviceResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the signature device
        algorithm:
          type: string
          description: Signature algorithm used by the device
          enum:
            - RSA
            - ECC
        label:
          type: string
          description: Human-readable label for the device (optional)
        signature_counter:
          type: integer
          description: Number of signatures generated by the device
    SignatureResponse:
      type: object
      properties:
        signature:
          type: string
          description: Base64 encoded signature
        signed_data:
          type: string
          description: Signed data
    SignatureRequest:
      type: object
      properties:
        data:
          type: string
          description: Data to be signed
    Error:
      type: object
      properties:
        errors:
          type: array
          items:
            type: string
                